agent_discovery/nanda/adapter/requirements.txt
flask
anthropic
requests
python-a2a==0.5.6
mcp
python-dotenv
flask-cors
pymongo

# Example: LangChain + Ollama
langchain-core
langchain-anthropic
langchain-ollama


# NANDA Adapter
nanda-adapter

# Example: MCP server
starlette
uvicorn
